fetching in a separate custom hook
