+++ frontend/src/components/contactView/ContactView.tsx
@@ -1,6 +1,4 @@
-import { useEffect, useState } from "react";
-import { contactsApi } from "../../api/contactsApi";
-import type { Contact } from "../../types/contact";
+import { useGetAllContacts } from "../../hooks/fetching/useGetAllContacts";
 import type { AppStateProps } from "../../types/state";
 import { ContactDetail } from "./ContactDetail";
 import { ContactList } from "./ContactList";
@@ -10,37 +8,7 @@ export default function ContactView({
   setSelectedContact,
   setCurrentPage,
 }: AppStateProps) {
-  const [contacts, setContacts] = useState<Contact[]>([]);
-  const [error, setError] = useState<string | null>(null);
-  const [isFetching, setIsFetching] = useState<boolean>(false);
-
-  useEffect(() => {
-    let isActive = true;
-
-    const loadContacts = async () => {
-      try {
-        setIsFetching(true);
-        const data = await contactsApi.getAllContacts();
-        if (isActive) {
-          setContacts(data);
-        }
-      } catch {
-        if (isActive) {
-          setError("Nepodařilo se načíst data.");
-        }
-      } finally {
-        if (isActive) {
-          setIsFetching(false);
-        }
-      }
-    };
-
-    loadContacts();
-
-    return () => {
-      isActive = false;
-    };
-  }, []);
+  const { contacts, setContacts, error, isFetching } = useGetAllContacts();
 
   return (
     <div className="list-view">
